+++ src/app/app.module.ts
@@ -1,4 +1,4 @@
-import { NgModule } from '@angular/core';
+import { InjectionToken, NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
@@ -6,11 +6,19 @@ import { ProductsModule } from './products/products.module';
 import { CartModule } from './cart/cart.module';
 import { SharedModule } from './shared/shared.module';
 import { OrdersModule } from './orders/orders.module';
+import { FirstComponent } from './first/first.component';
+import { ConstantService } from './core/services/constants.service';
+import { GENERATED_STRING, GeneratorFactory } from './core/services/generator.factory';
+import { GeneratorService } from './core/services/generator.service';
+import { APP_INITIALIZER } from '@angular/core';
+import { LocalStorageService } from './core/services/local-storage.service';
+import { ApplyStylesDirective } from './shared/directives/apply-styles.directive';
 
 
 @NgModule({
   declarations: [
-    AppComponent
+    AppComponent,
+    FirstComponent
   ],
   imports: [
     BrowserModule,
@@ -20,7 +28,25 @@ import { OrdersModule } from './orders/orders.module';
     OrdersModule,
     SharedModule
   ],
-  providers: [],
+  providers: [{
+    provide: ConstantService,
+    useValue: {
+      App: 'TaskManager',
+      Ver: '1.0',
+      API_URL: 'http://example.com/api'
+    }
+  },
+  GeneratorService,
+    {
+      provide: GENERATED_STRING,
+      useFactory: (generatorService: GeneratorService) => GeneratorFactory(10, generatorService),
+      deps: [GeneratorService],
+    },
+  {
+    provide: LocalStorageService,
+    useValue: new LocalStorageService(),
+  }
+  ],
   bootstrap: [AppComponent]
 })
 export class AppModule { }